Update readme to show the commands in the browser
As of now the readme page on the github doesn't show the
commands due to a formatting issue. This has been fixed.

The requirement for the bugzilla package hasn't been
mentioned anywhere. The script doesn't work without it and
the error message isn't good enough to ask the user to install it.
Hence have added it as a prerequisite in the readme.

Signed-off-by: hari gowtham <hgowtham@redhat.com>